Refactored ItemService and Controller

+++ src/main/java/com/siemens/internship/ItemController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 
 @RestController
 @RequestMapping("/api/items")
@@ -25,16 +26,19 @@ public class ItemController {
     @PostMapping
     public ResponseEntity<Item> createItem(@Valid @RequestBody Item item, BindingResult result) {
         if (result.hasErrors()) {
-            return new ResponseEntity<>(null, HttpStatus.CREATED);
+            //Return 400 (BAD Request) if validation fails
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity<>(itemService.save(item), HttpStatus.BAD_REQUEST);
+        //Return 201 otherwise
+        return new ResponseEntity<>(itemService.save(item), HttpStatus.CREATED);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Item> getItemById(@PathVariable Long id) {
         return itemService.findById(id)
                 .map(item -> new ResponseEntity<>(item, HttpStatus.OK))
-                .orElse(new ResponseEntity<>(HttpStatus.NO_CONTENT));
+                .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        //I modified from NO_Content to NOT_Found if the id doesn't exist in our db
     }
 
     @PutMapping("/{id}")
@@ -42,20 +46,27 @@ public class ItemController {
         Optional<Item> existingItem = itemService.findById(id);
         if (existingItem.isPresent()) {
             item.setId(id);
-            return new ResponseEntity<>(itemService.save(item), HttpStatus.CREATED);
+            //I changed from the CREATED status which is used for "CREATE" to OK which is suitable for UPDATE
+            return new ResponseEntity<>(itemService.save(item), HttpStatus.OK);
         } else {
-            return new ResponseEntity<>(HttpStatus.ACCEPTED);
+            //If the user is not found we need to return 404(NOT_FOUND)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteItem(@PathVariable Long id) {
-        itemService.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.CONFLICT);
+        //Check if the item exists before we try to delete it
+        if (itemService.findById(id).isPresent()) {
+            itemService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);  //204(successfully deleted)
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);  //404(item not found)
+        }
     }
 
     @GetMapping("/process")
-    public ResponseEntity<List<Item>> processItems() {
+    public ResponseEntity<CompletableFuture<List<Item>>> processItems() {
         return new ResponseEntity<>(itemService.processItemsAsync(), HttpStatus.OK);
     }
 }
